refactor(component): rename LazyExample component to HydrationButton
